+++ VlasovMS/03_lab_tables/sln/lab/lab.vcxproj.filters
@@ -57,5 +57,8 @@
     <ClInclude Include="..\..\include\TChain.h">
       <Filter>include</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\include\TChainNode.h">
+      <Filter>include</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
